rockspec: sync rockspecs with rocks.tarantool.org
Make lulpeg rockspecs, which are stored there, consistent with
rocks.tarantool.org ones. I don't know, whether there is a reason to
store rockspecs in the repository, but I see that we do. So let them be
in-sync with ones on the server.

See the relevant changes in tarantool/rocks:

* https://github.com/tarantool/rocks/commit/030c1e04257de0c9ea99f33b16090b7c06bc4b91
* https://github.com/tarantool/rocks/commit/749f0905a825af2ba2bd72dcf3493dcbb8b512a1
* https://github.com/tarantool/rocks/commit/a7013705226c830d3112b3900fda30e6ab9c3e00

After this, the rocks server goes to
https://github.com/tarantool/rocks.tarantool.org and rockspec changes on
the server are not tracked in git anymore (usually updated in
repositories and published to the server from CI).

After the transition we changed `git://` URLs to `git+https://`, see
https://github.blog/2021-09-01-improving-git-protocol-security-github/
for reasons.

Part of https://github.com/tarantool/tarantool/issues/6587